improve status bar layout
We had been truncating the right status bar text for years in a fashion that was hard to notice.  This has been more apparent on high DPI systems where the unscaled text was being visibly clipped.  We had previously adjusted that text for scaling and to avoid any clipping.  But the varied text lengths cause the shorter strings to display clunkily: this change updates the code to swap to a shorter scaled length once we switch away from the (longer) AppName boilerplate.